userservice/requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
requests==2.27.1
